# Allocate KV cache in contiguous memory for HPU performance