@mandy-li 2023-09-12
@puneeshkhanna , please provide performance measurement before/after this PR.