@puneeshkhanna 2023-09-13
@regisss  - thanks for merging. There will be very good performance boost with batch size scaling for all llama variants. Please check those cases too whenever you get a chance. 

Some numbers of 7B and 13B on 1x as below:
Llama 7B 
             baseline-> KV cache fix   -> KV cache fix + attn softmax BF16
BS-32   2460 -> 3137 tokens/sec -> 3243.76 tokens/sec
BS-64   3382 -> 4801 tokens/sec -> 5112 tokens/sec

Llama 13B
             baseline->  KV cache fix  -> KV cache fix + attn softmax BF16
BS-32   1416      -> 1760 tokens/sec -> 1812.85 tokens/sec

Please also review/merge https://github.com/huggingface/optimum-habana/pull/396.