@puneeshkhanna 2023-09-13
@regisss  - Just to add further, llama2 70b with 8x will also show considerable improvements with batching. Btw the fix which I did in repeat kv, maybe applicable generally to actual repeat kv code in transformers. What do you think ? I can push there too. If approved there, we can remove from here later.